requirements.txt
pycryptodome
aiogram
mysql-connector-python